refactor: inline @towles/shared into src/lib
`@towles/shared` was a private workspace package, never published to npm,
so every external `bun install @towles/tool` failed to resolve it. Bundling
it into the tarball doesn't work either: npm re-injects the bundled dep into
the registry packument as `@towles/shared: *`, which bun then 404s on.

It was only ever used by the root CLI (30 imports, none in agentboard), so
inline it as ordinary internal modules:

- Move packages/shared/src -> src/lib (renamed from "shared" since it is no
  longer a shared package).
- Rewrite the 30 bare `@towles/shared` imports to relative paths.
- Drop the dependency, bundledDependencies, the tsconfig path alias, and the
  packages/shared workspace.

The published package now has zero `@towles/*` dependencies and installs
cleanly. pack-publish.ts keeps stripping the root-only patchedDependencies.
