+++ src/commands/journal/paths.ts
@@ -1,7 +1,7 @@
 import path from "node:path";
 import consola from "consola";
 import { DateTime } from "luxon";
-import { getMondayOfWeek } from "@towles/shared";
+import { getMondayOfWeek } from "../../lib/index.js";
 import type { JournalSettings } from "../../config/settings.js";
 import { JOURNAL_TYPES } from "../../types/journal.js";
 import type { JournalType } from "../../types/journal.js";
