+++ src/lib/errors.test.ts
