+++ tsconfig.json
@@ -11,11 +11,8 @@
     "esModuleInterop": true,
     "skipDefaultLibCheck": true,
     "skipLibCheck": true,
-    "types": ["node", "@types/bun"],
-    "paths": {
-      "@towles/shared": ["./packages/shared/src/index.ts"]
-    }
+    "types": ["node", "@types/bun"]
   },
-  "include": ["src/**/*.ts", "bin/**/*.ts", "scripts/**/*.ts", "packages/shared/src/**/*.ts"],
+  "include": ["src/**/*.ts", "bin/**/*.ts", "scripts/**/*.ts"],
   "exclude": ["packages/agentboard"]
 }
